fix(adc): adc_raw_to_millivolts 64-bit buf  - Increase the size of the buffer used to convert a raw reading to millivolts. With higher resolutin ADCs (such as 24-bit) the 32bit buffer can result in overflow.

+++ include/zephyr/drivers/adc.h
@@ -985,7 +985,7 @@ static inline int adc_raw_to_millivolts(int32_t ref_mv,
 					uint8_t resolution,
 					int32_t *valp)
 {
-	int32_t adc_mv = *valp * ref_mv;
+	int64_t adc_mv = (int64_t)*valp * (int64_t)ref_mv;
 	int ret = adc_gain_invert(gain, &adc_mv);
 
 	if (ret == 0) {
